cdpr_mujoco/openvla-oft-requirements.txt
absl-py==2.3.1
accelerate==1.11.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
array_record==0.8.1
astunparse==1.6.3
attrs==25.4.0
bddl==3.6.0
cachetools==6.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.3.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
diffusers==0.30.3
dlimp @ git+https://github.com/moojink/dlimp_openvla@040105d256bd28866cc6620621a3d5f7b6b91b46
dm-tree==0.1.9
draccus==0.8.0
easydict==1.13
einops==0.6.1
etils==1.13.0
evdev-binary==1.9.2
exceptiongroup==1.3.0
fastapi==0.120.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.10.0
google-auth==2.42.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.2
json-numpy==2.1.1
jsonlines==4.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupytext==1.18.1
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
-e git+ssh://git@github.com/damurka5/LIBERO.git@8f1084e3132a39270c3a13ebe37270a43ece2a01#egg=libero
llvmlite==0.45.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mergedeep==1.3.4
ml-dtypes==0.2.0
mpmath==1.3.0
mujoco==3.3.7
mypy_extensions==1.1.0
nbformat==5.10.4
networkx==3.3
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opencv-python==4.11.0.86
OpenEXR==3.4.2
-e git+https://github.com/moojink/openvla-oft.git@e4287e94541f459edc4feabc4e181f537cd569a8#egg=openvla_oft
opt_einsum==3.4.0
packaging==25.0
peft==0.11.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
promise==2.3
protobuf==4.21.12
psutil==7.1.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyyaml-include==1.4.1
referencing==0.37.0
regex==2025.10.23
requests==2.28.1
requests-oauthlib==2.0.0
rich==14.2.0
robosuite==1.4.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.11.4
sentencepiece==0.1.99
sentry-sdk==2.43.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.49.1
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-addons==0.23.0
tensorflow-cpu==2.15.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.2
termcolor==3.2.0
tifffile==2025.5.10
timm==0.9.10
tokenizers==0.19.1
toml==0.10.2
tomli==2.3.0
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/moojink/transformers-openvla-oft.git@bc339d9ad707454c0c115970db43c260067c61ab
trimesh==4.9.0
triton==2.2.0
typeguard==2.13.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.13
uvicorn==0.38.0
wandb==0.22.3
Werkzeug==3.1.3
wrapt==1.14.2
zipp==3.23.0
